+++ omni/runtime/omni_rt.h
@@ -23,7 +23,11 @@
 // Basic I/O functions
 void omni_print_string(const char* str);
 void omni_println_string(const char* str);
+void omni_eprint_string(const char* str);
+void omni_eprintln_string(const char* str);
+void omni_io_flush(void);
 char* omni_read_line(void);
+char* omni_read_all(void);
 
 // Logging functions
 void omni_log_debug(const char* message);